+++ src/main/java/com/smartsheet/api/internal/SmartsheetImpl.java
@@ -22,6 +22,7 @@ package com.smartsheet.api.internal;
 
 
 import com.smartsheet.api.*;
+import com.smartsheet.api.internal.http.AndroidHttpClient;
 import com.smartsheet.api.internal.http.DefaultHttpClient;
 import com.smartsheet.api.internal.http.HttpClient;
 import com.smartsheet.api.internal.json.JacksonJsonSerializer;
@@ -249,6 +250,15 @@ public class SmartsheetImpl implements Smartsheet {
     private final AtomicReference<PassthroughResources> passthrough;
 
     /**
+     * Represents the AtomicReference for EventResources.
+     *
+     * It will be initialized in constructor and will not change afterwards. The underlying value will be initially set
+     * as null, and will be initialized to non-null at the first time it is accessed via corresponding getter, therefore
+     * effectively the underlying value is lazily created in a thread safe manner.
+     */
+    private final AtomicReference<EventResources> events;
+
+    /**
      * Create an instance with given server URI, HttpClient (optional) and JsonSerializer (optional)
      *
      * Exceptions: - IllegalArgumentException : if serverURI/version/accessToken is null/empty
@@ -301,6 +311,7 @@ public class SmartsheetImpl implements Smartsheet {
         this.imageUrls = new AtomicReference<ImageUrlResources>();
         this.webhooks = new AtomicReference<WebhookResources>();
         this.passthrough = new AtomicReference<PassthroughResources>();
+        this.events = new AtomicReference<EventResources>();
     }
 
     /**
@@ -426,6 +437,9 @@ public class SmartsheetImpl implements Smartsheet {
     public void setMaxRetryTimeMillis(long maxRetryTimeMillis) {
         if (this.httpClient instanceof DefaultHttpClient) {
             ((DefaultHttpClient) this.httpClient).setMaxRetryTimeMillis(maxRetryTimeMillis);
+        }
+        else if (this.httpClient instanceof AndroidHttpClient) {
+            ((AndroidHttpClient) this.httpClient).setMaxRetryTimeMillis(maxRetryTimeMillis);
         }
         else
             throw new UnsupportedOperationException("Invalid operation for class " + this.httpClient.getClass());
@@ -650,6 +664,18 @@ public class SmartsheetImpl implements Smartsheet {
             passthrough.compareAndSet(null, new PassthroughResourcesImpl(this));
         }
         return passthrough.get();
+    }
+
+    /**
+     * Returns the EventResources instance that provides access to events resources.
+     *
+     * @return the events resources
+     */
+    public EventResources eventResources() {
+        if (events.get() == null) {
+            events.compareAndSet(null, new EventResourcesImpl(this));
+        }
+        return events.get();
     }
 
     /**