+++ CHANGELOG.md
@@ -8,7 +8,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - isDisplaySummaryTasksEnabled is now named isDisplaySummaryTasks to fix serialization bug
 
-## [2.2.9] - 2019-02-5
+## [2.68.3] - 2019-06-27
+### Added
+- Initial support for an Android http client
+
+### Fixed
+- add CARD_DONE tag to supported list of column tags
+
+## [2.68.2] - 2019-06-17
+- Resolved Jackson security vulnerability
+
+## [2.68.1] - 2019-06-15
+- Travis build automation
+
+## [2.68.0] - 2019-05-09
+### Added
+- Implement Event Reporting
+
+## [2.2.9] - 2019-02-05
 ### Added
 - Added BASE URI definition for Smartsheetgov
 - Added group inclusion for GetCurrentUser